+++ Store.Api/Controllers/CategoryController.cs
@@ -9,7 +9,6 @@ namespace Store.Api.Controllers
     [Route("api/category")]
     public class CategoryController : ControllerBase
     {
-
         private readonly ICategoryService categoryService;
 
         public CategoryController(ICategoryService service)
@@ -44,45 +43,4 @@ namespace Store.Api.Controllers
             return Ok();
         }
     }
-
-    [ApiController]
-    [Route("api/brand")]
-    public class BrandController : ControllerBase
-    {
-
-        private readonly IBrandService brandService;
-
-        public BrandController(IBrandService brand)
-        {
-            brandService = brand;
-        }
-
-        [HttpGet]
-        public IActionResult Get()
-        {
-            return Ok(brandService.Get());
-        }
-
-        [HttpPost]
-        public IActionResult Post([FromBody] Brand brand)
-        {
-            brandService.Save(brand);
-            return Ok();
-        }
-
-        [HttpPut("{id}")]
-        public IActionResult Put(int id, [FromBody] Brand brand)
-        {
-            brandService.Update(id, brand);
-            return Ok();
-        }
-
-        [HttpDelete("{id}")]
-        public IActionResult Delete(int id)
-        {
-            brandService.Delete(id);
-            return Ok();
-        }
-    }
-
 }
