+++ Store.BD/Brand.cs
@@ -1,17 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Store.Db
+﻿namespace Store.Db
 {
     public class Brand
     {
-            public int Id { get; set; }
-            public string Name { get; set; }
-            public string Description { get; set; }
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public string Description { get; set; }
 
-            public virtual ICollection<Product>? Products { get; set; }
+        // foreigns    
+        public virtual ICollection<Product>? Products { get; set; }
     }
 }
