+++ Store.BD/Category.cs
@@ -6,6 +6,7 @@
         public string Name { get; set; }
         public string Description { get; set; }
 
+        // foreigns
         public virtual ICollection<Product>? Products { get; set; }
     }
 }
