+++ Store.BD/Product.cs
@@ -7,6 +7,7 @@
         public string Description { get; set; }
         public decimal Price { get; set; }
         public int BrandId { get; set; }
+        
         // foreigns
         public virtual ICollection<Category>? Categories { get; set; }
         public virtual Brand Brand { get; set; }
