+++ Store.BD/StoreDBContext.cs
@@ -13,22 +13,15 @@ namespace Store.Db
         {
             modelBuilder.Entity<Category>(Category =>
             {
-                Category.ToTable("Category");
-                Category.HasKey(p => p.Id);
-
                 Category.Property(p => p.Name).IsRequired().HasMaxLength(30);
                 Category.Property(p => p.Description).IsRequired(false).HasMaxLength(100);
             });
 
             modelBuilder.Entity<Brand>(Brand =>
             {
-                Brand.ToTable("Brand");
-                Brand.HasKey(p => p.Id);
-
                 Brand.Property(p => p.Name).IsRequired().HasMaxLength(30);
                 Brand.Property(p => p.Description).IsRequired(false).HasMaxLength(100);
             });
         }
-
     }
 }
