+++ Store.Services/ProductService.cs
@@ -49,7 +49,6 @@ namespace Store.Services
 
                 _context.SaveChanges();
             }
-
         }
     }
